Move commands to package
